rename repository files
